Fix getEClassifier(...) in the core lib and for static metamodels
The getEClassifier function for static metamodels use introspection
for registering all the defined classes. The current version uses
"global()" to get a full dictonnary of the registered components, but
it gives too much information and computes a bad subsets (called at
bad time) when the core project (pyecore.ecore) is imported inside a
static metamodel definition.

This commit introduces a new computation of this dictionnary using
methods from the introspect module. The computation also get rid of
'false positive' (i.e, classes that are indeed metaclasses but from an
imported module).