Fix issue on eOpposite for dynamic EClass instances
When a dynamic EClass is created with an eOpposite and one of the side
of the relation is set, the reflexive part tries to access the
opposite to set it. In dynamic mode, attributes are create on-the-fly
regarding the metamodel and if the opposite feature has not been
access first, it does not exists in the instance namespace leading to
an exception.

To avoid this behavior at some places, the default "setter" tries to
load the data first to force it's inclusion in the object namespace.