Fix issue on eContainer feature
When a feature is containment and does not own eOpposite reference,
then the value should reset it's container reference. This reference
was not properly updated and cause the 'eContainer()' feature to
return bad values.

This fix keep track of the old values to set it's container to None if
the object is removed from the EReferences (either set to None or
remove from the collection).