Add 'eStructuralFeatures' and "remove" eAttributes/eReferences
The use of eAttributes and eReferences was seducing in the first
place, but for XMI import/export, this will introduces complex
behavior to code. The big issue comes where an XMI path is given using
this syntax: //@feat.x/feat.y

In those cases, '.x' represents a position inside the features added
to the 'feat' meta-relationship. It means that the order is important
otherwise, it could lead to a bad meta-relationship resolving.

This commit brings back the 'eAttributes' feature while temporarily
removing the 'eAttributes/eReferences' feature from the Ecore
metamodel. Their behavior are coded inside '@property' functions at
the moment.