Add simple notion of 'decoder'
A 'decoder' is able to resolve uri-fragments. It reacts if it is able
to handle the uri and is able to resolve an uri using the
'resolve(...)' function.

Each 'Resource' (e.g: XMIResource) will also own a 'resolve(...)'
method so if a file resolver will be handled, it will be able to
return the ressource associated to a file as decoder.