Add model deserialization experiment for XMI
This commit proposes a new experiments where a simple metamodel had
been created in Eclipse (My.ecore) and a dynamic instance from this
metamodel (MyRoot.xmi). The metamodel is loaded from the XMI and added
into the 'global_registry'. Then, the instance model is loaded. On
this small example, the result is quite good, the instance model is
loaded regarding the metamodel previously loaded. There is a
cohabitation of static/dynamic instances wihout so much issues.

In this commit, the URI notion had also been introduced, but the
relations between ResourceSet/Resource and URI must be refined.

This experiment needs more work in order to be usable but the work
seems on the right track.