Add ECollection/ESet/EList/EOrderedSet and missing attributes
This commit introduces the principle of ECollections regarding
informations about the feature it represents. The current binding is
the following:

ordered/unique -> EOrderedSet
!ordered/unique -> ESet (set)
!ordered/!unique -> ???
ordered/!unique -> EList (list)

The only missing type is for the non ordered and non unique
collections. This will be also extended by frozenset and tuples in the
future to tackle the 'non changeable' property.