Fix issues on eAllOperations/eAllStructuralFeatures methods
The issue was in the computing of the set of
features/operations. These methods are called a lot during
serialization/deserialization and the current version is modifying the
object list each time instead of performing a shallow copy of the
collection.

The fix only introduce a copy of the list instead of sharing the list
instance.