Update the Ecore/UML test using ResourceSet
In this example, the UML.ecore is straight from the Eclipse
version. References are made to 'plateform:/plugin...' stuffs which
are related to Eclipse EMF plugin registered in memory. To deal with
this, this version registered in the ResourceSet (could also be done
in the global_registry) the missing resources.