Add shortucts methods and 'util' ones (eGet/eSet...)
This commit adds the 'eGet/eSet' methods on eobjects. These methods
are only shortcuts to express: '__getattribute__(name)' or
'__setattr__(name)'. They are able to work with string parameters or
'EStructuralFeature' ones.

Also, addition of the 'eContents' property which returns a list of the
objects directly contained in another. From this method a new method
'eAllContents()' add also been introduced, returning an iterator on
all the objects contained directly or by transitivity inside one.