Add first version of the 'save()' method on a resource
This methods currently only allows to save data in a file (IO will be
probably used in the future). The resource can be created empty and
elements added to it, using the 'append()' method directly on the
resource, or got from a previous one. Once elements have been added to
the resource, it can be saved using 'save()' or
'save(output="myfile.xmi")'. This method will create the file if the
resource does not exist.