Fix issue on 'eClassifiers' for static metamodels
This commit fixes the fact that each EClassifier added to the
'eClassifiers' var in a static EPackage (aka, a Python module) was not
setting the opposite relationship. This was an issue for Ecore
metamodel serialization.

This fix also ease the static metamodel production, the 'eClassifier'
var and the 'getEClassifier' method are automatically added by
PyEcore. This was, it is not required anymore to manually create them.