Improve performances (a little bit)
Performances on huge metamodel XMI deserialization are not so
great. The main issue come from the getattr and the setattr on
EObject. Each setattr performs a 'findEStructuralFeature' which is
quite slow as the EObject and all its super types are scanned.

The 'findEStructuralFeature' method had been rewritten a little bit,
but a new way of handling values in the (set/get)attr should be
found. A solution could be to store a new object biding the value to
the handled feature directly.