Add new URI management
The URI are now responsible for the creation of an input and output
file-like object (input for 'load()' and output for 'save()').

This commit also introduced a new URI type which is an HttpURI. This
new URI is able to create a file-like object from an http resource for
XMI (ecore/xmi).