Modify the XMI serialization so it deal with external resources
External resources (as Ecore types or other types) can be defined in
various resources. References to them must be done using a special uri
build around the resource uri and the uri fragment of the object. In
order to ease the fragment production, a method had been added (like
in the EMF Java sources) on EObjects. This method is overriden by
ENamedElement so it provides a special fragment in some cases. This
method should also be overriden in EModelElement EAnnotation and will
be in future versions.