Remove check on type during EClass promotion
The type checking on EClass during 'promotion' could be an issue when
dealing with static metamodels with inner packages. At this moment,
the package to register is no longer an EObject type, but a Python
module. The duck typing avoid method/call issues for following
threatments, so at this point, type test could be avoided.