Fix issue on static metaclass init if super is called
The previous implementation of 'MetaEClass' that was re-setting
systematically the '_isset' attribute and other ones by calling
'_subinit()'. This call is mandatory if no 'super().__init__()' call
had been expressed in a static EClass definition. This commit detects
if a super call had been performed or not and init accordingly the
static metaclass.