Add special behavior for static eclass serialization
When EClass are registered (dynamic one), the XMI serializer asks the
EObject to give it its uri fragment. In the case of static EClass,
this behavior cannot be got. However, each static EClass owns an
eClass attribute which references the dynamic EClass representation of
the static EClass. The fix is thus simple, when we try to serialize a
static EClass, we switch to it's dynamic EClass representation
instead.