Add first implementation of a basic notification system
A new type 'Observer' had been introduced, which can observe an
existing EObject. When it observes the eobject, it is notified each
time a modification is performed on the observed
element. Modifications are related to meta-property modifications as
ADD, REMOVE...

In order to make all the EObject observable, EObject now extends an
Observable class. This class sketches the basic behavior for
observable/observation.