Change ESet to EOrderedSet in every 'unique' case
The semantic of a property 'unique' and 'non ordered' fits the ESet
one. However, in the XMI serialization, the fact that relationship to
these elements is done by an indexed path is problematic. To solve
this, the idea is to use an EOrderedSet each time. The semantic is not
strictly the same, but references can be made to the object in XMI
using indexes.