Add a new way of searching elements by fragment
This new algorithm search in the directly contained elements. First,
some research are done by considering that the object can be an
EPackage, but then, a simple search on 'eContent' is done. This
algorithm does not give so poor performances. However, this could be
improve by another managment of EAttributes (using a kind of wrapper,
exactly as for the ECollections).