Add special distinction in XMI deserialization
A special split is now performed by searching information at two
places, for "href" element, information is searched in the pointed
resource while when there is a simple "type uri#fragment" path, the
information is searched inside the metamodel register instead. Not
sure this is the XMI format Eclipse deals about, but it seems to work.