Fix issue in cross-document XMI serialization
When a cross-document relationship is used, a 'href' tag must be
produced. Currently, the serializer asks the object to serialize if it
is part of the beeing serialized resource. If it is contained in
another, a 'href' tag is produced. Unfortunately, it was missing the
resource assignement to the root of an object placed in a
resource. Consequently, the serializer was not founding Resource for
the model root and thus was not able to serialize it.

The fix is very simple, once a model root is attached to a resource,
the model root eResource value is set.