Add 'name' as default acceptable fragment for XMI path
Previous algorithm to compute URI fragment for each EObject was
relying only on '@feature_name.index'. This new version also take into
account the fact that an EObject can have a 'name' feature. This
current algorithm has a flaw as it only check if the 'name' attribute
exists for the EObject.

However, if a user enter a 'name' attribute in the instance and if
'name' is not a EAttribute of it's eclass, the XMI serialization will
use the 'name' even if it should not.