Add strong coupling between EClass instances and Python class
EClass instances are already linked to Python class. This class is
used to create instances from the EClass instance. This commit
introduces a strong coupling by observing the EClass instance and
adding new type inheritance if the EClass eSuperTypes feature is
modified. This allows to keep a consistent super hierarchie regarding
the expressed metamodel.

In Python, this latter addition could be put aside because of the
duck-typing. However, this strong coupling and inheritance structure
in the Python class allows the user to dynamically had new functions
on super classes and let the subclasses instances use them.