Add restriction on inheritance update for static metaclasses
In the context of static metaclasses, the fact that the inheritance
tree is recomputed each time during the class decoding/reading in the
Python module is a huge waste of time. At the moment, the introduced
line avoid static classes inheritance to be modified even after the
MetaEClass "execution". It could be changed in the future as a was of
only avoiding it during MetaEClass initialization in order to allows
dynamic inheritance tree modification on static metaclasses.