Fix issue with default value initiatlization for static metaclasses
When static metaclasses instances are created, the metainitializer
tries to set the default value for each element in the created
instance. This version was containing a bug that didn't properly set
the default value of EAttributes. As a result, the default value was
the EAttribute and not the default value of this EAttribute.

This commit fixes this by adding default type in the
metainitialization and by adding some test cases for this issue. In
the same time, the order of the metainitiliazation had been
reversed. First it was init attribute from child to parent, now it is
from parent to child. The reason is that if a child 'masks' a property
and change it's name, this is the masked type that should be the
default property type, not the parent one.