Fix issue about eResource not updated when element is added to a Resource
When an element is added into a Resource, all the elements should
shift resource also. This is already done by the 'append' method on a
Resource. However, when an element is added to another one in a
resource (by a containment reference), the added element should adopt
the container resource. This commit fix this latter behavior as it
introduces errors during XMI serialization.