Improve performances a little bit
The big struggle is when a simple 'obj.att = xxx' is used, at the
moment, this pass through the big settattr that performs all the
checks and tries to do the famous 'findEStructuralFeature'. This could
be solved if a dedicated class is used for representing the data for
EAttribute. This will be a major improvement over performances.