Add eternal listener concept
Currently, there is some operations on EClass that are performed by
notfications. A special observer is added to create EClass
instances. As these listeners are in the same 'listener' list than
possibly added ones, they could be erased by mistake. An eternal
listeners is a listeners that cannot be removed.