Add to_string for EBoolean
The to_string method is called during XMI serialization to provide an
string representation of the in-memory value. Boolean values are
gathered as 'True' value in Python while in Java, they are 'true'
(lower case). The current version misses a 'to_string' for EBoolean
and, in the same row, a fix (typo) had been fixed in the EDataType
class.