Remove resource-less objects from XMI serialization
This is a first step towards objects removal. The added behavior
allows the user to "remove" elements in a way. If an element is not
contained in a resource anymore, the reference towards the object is
not serialized. This way, anytime an object is removed from a
container and let 'in the void', XMI serialization will get rid of
it. However, this new addition requires that the Ecore metamodel is
always loaded in the global_registry (in case someone wants to
serialize ecore files) as a metamodel can references basic types
(EString, EBoolean) which are basically not contained in a resource.