Add serialization for proxy elements
Proxy serialization is a little bit special as it requires that the
element must be stored as an href in the xmi. In this case, the logic
is simple, if the proxy is not resolved, then an href with the
recorder path towards the element is created. If the proxy had been
resolved, the path is computed as if it were an existing object (in
case the object had moved from a resource to another).