Add a 'normalize' method on URI
This normalization method will be used for resolving the full path of
files. The normalized full path will be the key registered in the
resource dict used by the resource set. This will be used for XMI
deserialization (normalization of relative path) and XMI serialization
(relative path regarding an existing resource). This should give
enough flexibility for cross-references resolving.