Add proxy support for loading/saving
This commit introduces the special behavior that occurs when resources
are saved/loaded while using EProxy. During load, the uri is
normalized in order to look it in the resource set and during save,
the normalized uri is used to compute the relative path from one
resource to the other.