Fix double resources loading in same ResourceSet
With the current code, two 'get_resource()' call of the same URI in
the same ResourceSet (i.e: the resource had already been loaded)
returns two different resources. This commit fix this: if a resource
had already been loaded in a ResourceSet, a second call of the
'get_resource' returns the previously created resource. This fix is
critical for the next feature (object deletion).