Add first version of the delete method on the EObjects
The delete method has a special behavior regarding if the method is
called from a proxy or from an EObject. Each time the method is
called, it searched if the element contained is the collection is a
proxy and, if the proxy is resolved, if the pointed element is the one
to delete. The idea, for the method, is to clean all the collections
as the best as it can.