Add performance improvement
This small fix for performances takes place in the '_initmettaattr'
method. This method init all the meta attribute by setting them to the
default value. The current call is quite time consumming as it always
tries to put the opposite relationship in the same time. At this step
of the object creation, it is not yet relevant.