Add delete feature for EObject and EProxy
The delete feature allows the user to delete a part of a model. The
part can be a simple element or a sub-graph. The delete tries to keep
up to date a special list that gathers the non-inverse navigable
relation. When called, the method gathers all the EReferences of the
object to delete and these special relations. It then update the
pointed references. There is a special behavior if the object to
delete is a proxy. If unresolved, the proxy can only be removed from
the main location, but not from the remote one. If resolved, the proxy
keep the classical behavior. This behavior tries to match the EMF-Java
one:

https://www.eclipse.org/forums/index.php/t/127567/