Add better enumeration for Notification
The previous 'Kind' enumeration for Notification were using a
home-made-cooked way of dealing with enumeration. The code were got
from an article from the http://sametmax.com/ website (great website
by the way). This new version uses the python 3.4 enumeration module,
but as this module had been introduced in Python 3.4, it means that
the Python 3.3 compatibility is broken.