Add explicit dict and function in the generated module
The previous version of the generator was 'hidding' the creation of
the 'eClassifiers' dict and the 'getEClassfier()' function. These
elements were dynamically created by PyEcore and added inside the
generated module namespace. This way of doing stuffs was a confusing
and was not the best way of dealing with them. This new version makes
these element creation explicit in the generated code.