Add new class to ease dynamic metamodel handling
The dynamic metamodel manipulation is a little bit cumbersome when it
comes to extract all the existing EClass from a loaded
EPackage. Currently, a solution is to gather all the elements by hand
in dedicated variables:

>>> A = dynamic_epackage.getEClassifier('A')
>>> B = dynamic_epackage.getEClassifier('B')
>>> a1 = A()
>>> b1 = B()

This commit introduces a new class: 'DynamicEPackage' which
constructs, using reflection, an object that has direct references to
each EClass/sub-EPackage by name:

>>> from pyecore.utils import DynamicEPackage
>>> mylib = DynamicEPackage(dynamic_epackage)
>>> a1 = mylib.A()
>>> b1 = mylib.B()

This greatly helps the user to easily call and get EClass from a
freshly loaded dynamic EPackage.