Fix double notification raised for eopposite references
When an eopposite reference were set, the notification system were
called three times: one for the main feature (the feature on which the
add/remove/set/unset have been made by the user) and two for the
eopposite. The first eopposite notification were normal, but the
second one was a residual notification sent by the algorithm. This new
commit simply removes the extra-notifications and adds new tests to
detect these issues.