Add better semantic for some type of collections
The imlementation of some collection type cannot be done easily
because of the XMI import/export. The XMI import/export requires that
the elements can be 'indexed'. The only way of ensuring this is to
rely on 'non-set' imlementation, which is weired for 'unique' and/or
some 'unordered' collections. This commit still uses the classical
EList and EOrderedSet (that are indexable), but also proposes two new
types 'EBag' and 'ESet' which, at the moment, inherits from 'EList'
and 'EOrderedSet'. At least, this commit introduce a better semantic
for each collection.