Fix issue with Ecore inheritance when loading '.ecore'
In '.ecore' definition, EClass can inherits from Ecore basic EClasses,
e.g: an EClass named 'Foo' can inherits from 'EModelElement' in order
to gain access to 'EAnnotations' supports. This kind of handling is
not a classical flow, but it can occur. This fix tries to 'harmonizes'
the way the python classes and the EClasses are handled so this
inheritance is not considered an issue anymore.