Fix 'prepare' issue for Remove command
The 'can_execute' property on each command prepare the command and
sets some of the internal attributes of the commands while checking
the command sanity. The previous command check was a little bit to
harsh and wasn't taking into account the fact that the value could be
ommited if an index was set.