Fix issue with computation of inverse relations
When an inverse relation has already been set by a relation (for
single relation) and a new 'eSet' was performed with the same object
in the same relation:

    >>> a.eSet(rel) = b
    >>> a.eSet(rel) = b

the current algorithm was deleting the old record of the inverse
relationship set from the the previous value after it set the new
record in the new value. In case the old and the new value is the
same, the removal was introducing an issue as it flushed out the
perfectly good record. As simple fix, the record removal is now
performed before the addition.