Fix shared content for mutable EDatatype
When mutable EDatataypes are defined (e.g: EStringToStringMapEntry),
each default value was pointing to the same shared value (exactly the
same thing that when 'def x(self, n={})'). The default_value is now
computed, if a special attribute is set, the default value is always
created as a new empty value.