Fix default value for EAttribute
EAttribute let the ability to express default values. This value is
assigned when an EClass instance is created. The default_value is
computed as follow: if the EAttribute default_value is set, this
default_value is returned. If the default_value of the Eattribute is
not set, then the default_value of the EAttribute associated EDataType
is set. This way of computing elements was not properly used during
instance initialization.