Fix decoding issue when HttpURI with http-href is used
When a href is used, the ResourceSet resolver tries to concatenate the
path built from the main uri resource and the href uri fragment. In
the case of HttpURI, the concatenation provided a
'http://abc/http://cde' like uri. The 'normalize()' method of URI was
spliting on '://' and used unpacking to two vars exactly. With this
kind of uri, it resulted in an exception. This commit fixes this issue
using simply the 'maxsplit' option from the 'split()' method.