Fix intra-document references by proxy
A reference between elements can also be done using a 'full' URI, i.e:
specifying the uri/path of the resource to access and the path towards
the object. This way of referencing elements is not reserved to
metamodel references, but can be done with any kind of references. To
deal with this, a proxy is introduced each time such a reference is
done. This allows to relies on the same mechanism as the href one and
gives a better control over their resolutions.