Fix ResourceSet local resource resolving
When a local resource is searched, the path and its uri is split. Once
the uri is split, its path is searched in the 'resources' of the
ResourceSet. This search was done in a 'file' like only researched,
while the uri could be a logical one (for the 'plateform:/' like uri).