Fix missing 'name' feature validation
The name feature was only handled as a simple python attribute instead
of an EAttribute. This time, the 'name' feature is handled as an
EAttribute. As each instance of EAttribute needs to use its own name
(which is an EAttribute), it is required to cut the recursive call. To
do so, the EStructuralFeature listen to each changes performed on
itself. If a modification occurs on the 'name' feature, it keeps a
simple python attribute version which can be used in the
EStructuralFeature descriptor.