Fix resource creation if model loading fails for ResourceSet
Currently, even if the model loading fails, a resource is created
in a ResourceSet. This behavior was problematic as two successive
loading of the same faulty model lead to a failure the first
time, but to a success the next time. This commit fixes this
behavior and adds more tests to detect possible regressions about
this.