Improve OrderedSet 'insert/pop' methods and add basic __setitem__
The default OrderedSet implementation does not provide methods
for 'insert' and 'pop'. The current code provided by PyEcore,
monkey patching the library, was mixed with internal PyEcore
code. This new implementation split the two concerns and proposes
a better way of dealing with these two methods. Also, a very
basic support for __setitem__ is provided. This basic
implementation relies on insert and pop to do the job.