Add 'load' method first version for JsonResource
The load method is used by the 'get_resource' to automatically get the
content of a file. Some basic tests are also part of this commit, they
test the JsonResource registration in the factory, a simple ecore file
in a Json format load and the loading of two files 'f1' and 'f2' with
cross-references from f1 to f2.  With the load method, a very small
refactoring of Resource have been made, pushing the 'resolve' method
from the XMIResource and JsonResource to 'Resource'. More tests are
to come.