Fix python < 3.6 compatibility for input stream
The json.load function from the json library takes a file descriptor
as parameter and reads in it. In Python 3.6, the file can be opened in
a binary mode, and everything's fine. However, for other Python 3
version, this does not work. The solution is to read first each bytes,
decode them as 'utf-8' and finally load them using json.loads instead
of json.load.