Skip EObject when static EClass is registered
When a static metaclass is used, the '@EMetaclass' annotation (or the
MetaEClass metaclass) iterates through all the class base classes to
register and keep up to date the 'eClass' representation of the Python
class. If the 'EObject' class is not skiped during the eClass
creation, the meta-object 'EObject' appears as super types. This
information is redundant and is not relevant for the eClass
representation as it is a technical detail.