Add dedicated method for eclass ref serialization in JsonResource
This method allows the user to override it if required in order to
control a little bit the way the JSON will be produced. If the users
override this method by providing a new implementation, they also
should override the eclass resolve method. Otherwise, the
deserialization of a previously serialized model with this new
implementation will fail.

The two methods to override are the following:

- serialize_eclass(self, eclass)
- resolve_eclass(self, uri_eclass)

Note that for the 'resolve_eclass' method, the use of a cache like
'lru_cache' is often a good idea.