Add special deserialization method for EEnum
The basic deserialization method for EEnum instance was the same than
the one for EDataType. They only takes the string and put it in the
feature instance. Instead, when a string is found for an EEnum
feature, the EEnumLiteral must be searched. This new from_string
implementation just does this.