Add metamodel relative path auto-loading
When a metamodel makes a relative reference to another one, the other
metamodel needs to be loaded first. However, for scripting purposes,
this can be cumbersome, and complex. This new feature enables the
auto-loading of relative metamodel path when required. By "when
required", it means that an EProxy is added when a cross-reference is
detected, but if the EProxy must be resolved (because of a
navigation), if the associated Resource does not exists in the
ResourceSet, it is automatically loaded and resolved.