Add missing 'type' field for XMI href serialization
The 'xsi:type' field is required when a crossreference is
serialized. This field is used to check if the resolved resource has
the right type (not really, but it help). This field is kinda optional
at the moment for PyEcore, but required for Java-EMF, otherwise the
EMF will not be able to create the right proxy for the crossreferenced
resource.