Add first support for derived collections
The derived collections are obtained from many features that are also
marked as derived. A derived collection does not stores nor gets
direct information, instead it fetches information from other
collection, compute a result and store it. In the realisation, a
derived collection will inherits from one of the generic derived collections
that only proposed an empty implementation.