Add fist support for EGenerics
This implementation is currently incomplete as it is not possible to
express dedicated types for instances. This capability will be added in
future versions. The current version is able to catch only a part of the
generic semantic, but at least, it does not prevent from opening files
and models that uses them (it should not).