Update generic implementation with a kind of cache
This cache can also be used to give a specific type to a relation that
is typed by a ETypedParameter. Here is an example:

Let's assume we have a class templated by 'T' and a relation going from
another EClass to T (named 'testrelation'). You can define for one instance the constraint type
of the relation like this:

    x.testrelation
    x.__dict__['testrelation'].generic_type = mytype_constraint

This solution requires more implementation, tests and a better frontend
to be exploited in good conditions. This solution should also make use
of the 'eBounds' on a feature in order to ensure that the type placed as
"constraint" is compatible with what is expressed in the bounds of the
generic type.