Add support for item addition from an iterable
If an iterable (beside strings) is passed as redefinition of a collection
(eg: ``a.b = [2, 3]``), the previous behavior was to raise an error for
incompatible type. This new version just throws all the previous element
from the ``a.b`` collection and place all elements from the other one in it.
This behavior will introduce easier handling for model transformations.