Fix enumeration literal addition as attribute
In the case literals where added to an enumeration using the 'eLiterals' collection, this implied that the 'EEnum' instance was not updated with the required instance attribute. This new implementatio uses a listener to catch additions that are made to the enumeration in order to either add or remove instance attributes.